LayoutTests/platform/glib/fast/mediastream/RTCPeerConnection-setRemoteDescription-offer-expected.txt

Test RTCPeerConnection.setRemoteDescription called with an RTCSessionDescription of type 'offer'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


*** Create (remote) offer with audio (remoteOffer1)
*** Done, start testing with remoteOffer1
PASS pc.signalingState is 'stable'
PASS pc.remoteDescription is null
PASS pc.pendingRemoteDescription is null
PASS pc.currentRemoteDescription is null

*** 'track' event must fire before setRemoteDescription promise is fulfilled
*** Check track event properties
PASS pc.getReceivers().includes(trackEvent.receiver) is true
PASS pc.getTransceivers().includes(trackEvent.transceiver) is true
*** Remote track should be muted initially
PASS trackEvent.track.muted is true
*** Remote track properties should match track added at remote side (remotePc)
PASS trackEvent.track.kind is remotePcAudioTrack.kind
PASS trackEvent.track.id is remotePcAudioTrack.id
FAIL trackEvent.track.label should be remote audio. Was .
*** Check local and remote transceivers
PASS trackEvent.transceiver.mid is remotePcAudioTransceiver.mid

PASS remoteOffer1 set as remote description
PASS pc.signalingState is 'have-remote-offer'
FAIL pc.remoteDescription should be [object Object]. Was [object RTCSessionDescription].
FAIL pc.pendingRemoteDescription should be [object Object]. Was [object RTCSessionDescription].
PASS pc.currentRemoteDescription is null

PASS remoteOffer1 set as remote description (again)
PASS pc.signalingState is 'have-remote-offer'

*** Try setting local descriptions with bad types for the current state
PASS promise pc.setRemoteDescription({type:'answer', sdp:remoteOffer1.sdp}); rejected with InvalidStateError: Failed to set remote answer sdp
PASS promise pc.setRemoteDescription({type:'pranswer', sdp:remoteOffer1.sdp}); rejected with InvalidStateError: Failed to set remote answer sdp

*** Create (remote) offer with video (remoteOffer2)
*** Done, start testing with remoteOffer2
PASS pc.signalingState is 'have-remote-offer'

*** 'track' event must fire before setRemoteDescription promise is fulfilled
*** Check track event properties
PASS pc.getReceivers().includes(trackEvent.receiver) is true
PASS pc.getTransceivers().includes(trackEvent.transceiver) is true
*** Remote track should be muted initially
PASS trackEvent.track.muted is true
*** Remote track properties should match track added at remote side (remotePc)
PASS trackEvent.track.kind is remotePcVideoTrack.kind
PASS trackEvent.track.id is remotePcVideoTrack.id
FAIL trackEvent.track.label should be remote video. Was .
*** Check local and remote transceivers
PASS trackEvent.transceiver.mid is remotePcVideoTransceiver.mid

PASS remoteOffer2 set as remote description
PASS pc.signalingState is 'have-remote-offer'
FAIL pc.remoteDescription should be [object Object]. Was [object RTCSessionDescription].
FAIL pc.pendingRemoteDescription should be [object Object]. Was [object RTCSessionDescription].
PASS pc.currentRemoteDescription is null

PASS successfullyParsed is true

TEST COMPLETE
